RPCHandler implementation
